Extract json log reader service from web-api to separate dot net standard project to reuse it in a wpf project
