+++ LogViewer.Server.Tests/LogViewer.Server.Tests.csproj
@@ -13,6 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\LogViewer.JsonLogReader\LogViewer.JsonLogReader.csproj" />
     <ProjectReference Include="..\LogViewer.Server\LogViewer.Server.csproj" />
   </ItemGroup>
