+++ LogViewer.Server/LogViewer.Server.csproj
@@ -18,10 +18,11 @@
     <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="2.2.0" />
-    <PackageReference Include="Serilog.Filters.Expressions" Version="2.0.0" />
-    <PackageReference Include="Serilog.Formatting.Compact.Reader" Version="1.0.3" />
-    <PackageReference Include="Serilog.Sinks.File" Version="4.0.0" />
     <PackageReference Include="Microsoft.Packaging.Tools.Trimming" Version="1.1.0-preview1-26619-01" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\LogViewer.JsonLogReader\LogViewer.JsonLogReader.csproj" />
   </ItemGroup>
 
 </Project>
